util functions refactor
